+++ tests/models/test_feed.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 from tests.cases import ModelTestCase
 from tests.factories import ShopFactory
 from yandex_market_language import models
@@ -19,4 +21,14 @@ class FeedModelTestCase(ModelTestCase):
         self.assertEqual(list(el.tag for el in feed_el), ["shop"])
         self.assertElementsEquals(feed_el[0], shop.to_xml())
         self.assertEqual(feed_el.tag, "yml_catalog")
-        self.assertEqual(feed_el.get("date"), feed.date)
+        self.assertEqual(feed_el.get("date"), feed._date)
+
+    @mock.patch("yandex_market_language.models.Shop.from_xml")
+    def test_from_xml(self, p):
+        shop = ShopFactory()
+        feed = models.Feed(shop)
+        feed_el = feed.to_xml()
+        p.return_value = shop
+        parsed_feed = models.Feed.from_xml(feed_el)
+        self.assertEqual(p.call_count, 1)
+        self.assertEqual(feed.to_dict(), parsed_feed.to_dict())