+++ yandex_market_language/models/price.py
@@ -1,12 +1,22 @@
+from typing import Optional
+
 from .base import BaseModel, XMLElement
 
 from yandex_market_language.exceptions import ValidationError
 
 
 class Price(BaseModel):
-    def __init__(self, value, is_starting: bool = False):
+    def __init__(self, value, is_starting=False):
         self.value = value
         self.is_starting = is_starting
+
+    @property
+    def is_starting(self) -> Optional[bool]:
+        return self._str_to_bool(self._is_starting)
+
+    @is_starting.setter
+    def is_starting(self, value):
+        self._is_starting = self._is_valid_bool(value, "is_starting", True)
 
     @property
     def value(self):
@@ -29,3 +39,7 @@ class Price(BaseModel):
             el.attrib["from"] = "true"
         el.text = self.value
         return el
+
+    @staticmethod
+    def from_xml(el: XMLElement) -> "BaseModel":
+        return Price(el.text, el.attrib.get("from", False))