+++ yandex_market_language/parser.py
@@ -1,14 +1,23 @@
+# from xml.dom import minidom
 from xml.etree import ElementTree as ET
 
-from yandex_market_language.models import Feed, Shop, Currency, Category
+from yandex_market_language.models import Feed
 
 
 class YMLParser:
     def __init__(self, file_or_path):
         self._file_or_path = file_or_path
-        self._tree = ET.parse(file_or_path)
+
+    # @staticmethod
+    # def prettify(el: ET.Element) -> str:
+    #     """
+    #     Return a pretty-printed XML string for the Element.
+    #     """
+    #     raw = ET.tostring(el, "utf-8")
+    #     parsed = minidom.parseString(raw)
+    #     return parsed.toprettyxml(indent="\t")
 
     def parse(self) -> Feed:
-        currencies = [Currency("RUB", "1")]
-        categories = [Category("123", "Test")]
-        return Feed(Shop("test", "", "", currencies, categories))
+        tree = ET.parse(self._file_or_path)
+        root = tree.getroot()
+        return Feed.from_xml(root)